Fix fetching an overriden package from Goopfile.lock when the original repo is gone
If there's a repo that no longer exists, the
vcs.RepoRootForImportPathStatic(..) will try to ping that repo and will
fail with an error (404), therefore the whole
RepoRootForImportPathWithURLOverride(..) function will fail, failing the
whole override mechanism.

So this is a bit smarter in the sense that it passes the override url to
vcs.RepoRootForImportPathStatic(..) and then returns whatever that
function actually returns, no need to replace the url there.